refactor: Split daemon/rpc.rs into focused sub-modules
Extract 4 cohesive modules from the monolithic 4466-line rpc.rs:
- rpc_kanban.rs: Kanban board data, GraphQL queries, PR parsing, CI status
- rpc_tasks.rs: Task CRUD (create/update/done/metadata/claim/request)
- rpc_sessions.rs: Session attach/detach/list for interactive debugging
- rpc_channel.rs: Channel post/read, status, reminders, mention routing

rpc.rs remains as the dispatch spine (~1650 lines) with coworker
lifecycle handlers, auth switch, and param extraction helpers.
